requirements.txt
Django==2.2.2
django-bootstrap4==0.0.6
django-crispy-forms==1.7.2
ipython==7.2.0
requests==2.21.0
djangorestframework==3.8.2
